+++ core/core-api/build.gradle
@@ -6,6 +6,7 @@ dependencies {
     implementation project(":integration:db-core")
     implementation project(":integration:plc-integration")
     implementation project(":support:logging")
+    implementation project(":integration:plc-api")
 
     implementation 'org.springframework.boot:spring-boot-starter-web'
     implementation 'org.springframework.boot:spring-boot-starter-websocket'
