+++ core/core-api/src/main/java/com/hbc/pms/core/api/support/data/DataProcessor.java
@@ -1,8 +1,6 @@
 package com.hbc.pms.core.api.support.data;
 
-import com.hbc.pms.plc.integration.huykka7.IoResponse;
-import com.hbc.pms.plc.integration.mokka7.exception.S7Exception;
-import com.hbc.pms.plc.io.Blueprint;
+import com.hbc.pms.plc.api.IoResponse;
 import org.springframework.stereotype.Component;
 
 import java.util.HashMap;
@@ -14,17 +12,11 @@ public class DataProcessor {
         Map<String, String> flattenedData = new HashMap<>();
 
         for (Map.Entry<String, IoResponse> entry : rawData.entrySet()) {
-            try {
-                var value = entry.getValue().getValue().toString();
-                if (Float.parseFloat(value) == 0) {
-                    value = "x";
-                }
-                flattenedData.put(entry.getKey(), value);
-            } catch (S7Exception e) {
-                if (e.getErrorCode() == -1) {
-                    flattenedData.put(entry.getKey(), "x");
-                }
+            var value = entry.getValue().getValue().toString();
+            if (Float.parseFloat(value) == 0) {
+                value = "x";
             }
+            flattenedData.put(entry.getKey(), value);
         }
         return flattenedData;
     }
