+++ integration/plc-api/src/main/java/com/hbc/pms/plc/api/DataType.java
@@ -1,4 +1,4 @@
-package com.hbc.pms.plc.integration.huykka7;
+package com.hbc.pms.plc.api;
 
 public enum DataType {
     BIT,
