+++ integration/plc-api/src/main/java/com/hbc/pms/plc/api/IoResponse.java
@@ -1,8 +1,5 @@
-package com.hbc.pms.plc.integration.huykka7;
+package com.hbc.pms.plc.api;
 
-import com.hbc.pms.plc.integration.mokka7.exception.S7Exception;
-import com.hbc.pms.plc.integration.mokka7.util.S7;
-import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
@@ -23,24 +20,14 @@ public class IoResponse {
   private byte[] rawData;
   private PlcValue plcValue;
 
-  public <T> T getValue() throws S7Exception {
+  public <T> T getValue() {
     Object value = null;
-    if (plcValue != null){
       switch (this.dateType){
         case BIT -> value = plcValue.getBoolean();
         case D_INTEGER -> value = plcValue.getInteger();
         case DOUBLE -> value = plcValue.getDouble();
-        default -> throw new S7Exception(-1, "Not supported data type");
+        default -> throw new RuntimeException("Not supported data type");
       }
       return (T)value;
-    }
-
-    switch (this.dateType){
-      case BIT -> value = S7.getBitAt(rawData[0],0);
-      case D_INTEGER -> value = S7.getDIntAt(rawData, 0);
-      case DOUBLE -> value = S7.getFloatAt(rawData, 0);
-      default -> throw new S7Exception(-1, "Not supported data type");
-    }
-    return (T) value;
   }
 }
