+++ integration/plc-integration/build.gradle
@@ -3,5 +3,5 @@ dependencies {
     implementation 'org.apache.plc4x:plc4j-api:0.11.0'
     implementation 'org.apache.plc4x:plc4j-driver-s7:0.11.0'
     implementation project(":core:core-model")
-
+    implementation project(":integration:plc-api")
 }
