--- integration/plc-integration/src/main/java/com/hbc/pms/plc/integration/huykka7/PlcCollector.java
@@ -1,2 +1,0 @@
-package com.hbc.pms.plc.integration.huykka7;
-
