--- integration/plc-integration/src/main/java/com/hbc/pms/plc/integration/huykka7/PlcConnectionConfiguration.java
@@ -1,15 +1,0 @@
-package com.hbc.pms.plc.integration.huykka7;
-
-import lombok.Builder;
-import lombok.Data;
-
-@Data
-@Builder
-public class PlcConnectionConfiguration {
-  private String ipAddress;
-  private int rack;
-  private int cpuMpiAddress;
-
-  @Builder.Default
-  private int port = 102;
-}
