--- integration/plc-integration/src/main/java/com/hbc/pms/plc/integration/mokka7/S7Config.java
@@ -1,88 +1,0 @@
-package com.hbc.pms.plc.integration.mokka7;
-
-import com.hbc.pms.plc.integration.mokka7.type.ConnectionType;
-import lombok.Getter;
-
-public class S7Config {
-    private ConnectionType type;
-    @Getter
-    private String host;
-    private int port;
-    private int slot;
-    private int rack;
-    private int remoteTSAP;
-    private int localTSAP;
-
-    public S7Config() {
-        localTSAP = 0x0100;
-        remoteTSAP = -1;
-        type = ConnectionType.PG;
-        port = 102;
-        slot = 0;
-        rack = 0;
-
-    }
-
-    public ConnectionType getType() {
-        return type;
-    }
-
-    /**
-     * Sets the connection resource type, i.e. the way in which the Clients connects to a PLC.
-     *
-     * @param type ConnectionType
-     */
-    public void setType(ConnectionType type) {
-        this.type = type;
-    }
-
-    public void setHost(String host) {
-        this.host = host;
-    }
-
-    public int getPort() {
-        return port;
-    }
-
-    public void setPort(int port) {
-        this.port = port;
-    }
-
-    public int getRemoteTSAP() {
-        if (remoteTSAP < 0) {
-            remoteTSAP = (type.getValue() << 8) + (rack * 0x20) + slot;
-        }
-        return remoteTSAP;
-    }
-
-    public void setRemoteTSAP(int remoteTSAP) {
-        this.remoteTSAP = remoteTSAP;
-    }
-
-    public int getLocalTSAP() {
-        return localTSAP;
-    }
-
-    public void setLocalTSAP(int localTSAP) {
-        this.localTSAP = localTSAP;
-    }
-
-    public int getSlot() {
-        return slot;
-    }
-
-    public void setSlot(int slot) {
-        this.slot = slot;
-    }
-
-    public int getRack() {
-        return rack;
-    }
-
-    public void setRack(int rack) {
-        this.rack = rack;
-    }
-
-
-
-}
