+++ settings.gradle
@@ -11,4 +11,7 @@ include 'core:core-api'
 include 'core:core-model'
 include 'integration:db-core'
 include 'integration:plc-integration'
+include 'integration:plc-api'
 include 'support:logging'
+
+
